Messed up.

+++ src/java/ch/usi/da/paxos/message/Message.java
@@ -38,7 +38,7 @@ public class Message implements Serializable {
 
 	private static final long serialVersionUID = -4938636847085992695L;
 	
-	private final int instance;
+	private final long instance;
 	
 	private final int sender;
 	
@@ -62,7 +62,7 @@ public class Message implements Serializable {
 	 * @param ballot the ballot number
 	 * @param value the value (can be null)
 	 */
-	public Message(int instance,int sender,PaxosRole receiver,MessageType type,int ballot,Value value){
+	public Message(long instance,int sender,PaxosRole receiver,MessageType type,int ballot,Value value){
 		this.instance = instance;
 		this.sender = sender;
 		this.receiver = receiver;
@@ -74,7 +74,7 @@ public class Message implements Serializable {
 	/**
 	 * @return the instance
 	 */
-	public int getInstance() {
+	public long getInstance() {
 		return instance;
 	}
 
@@ -235,7 +235,7 @@ public class Message implements Serializable {
 	public static long getCRC32(Message m){
 		if(m == null) return 0;
 		CRC32 crc = new CRC32();
-		crc.update(m.getInstance());
+		crc.update((int) m.getInstance());
 		crc.update(m.getSender());
 		crc.update(m.getReceiver().getId());
 		crc.update(m.getType().getId());
@@ -261,7 +261,7 @@ public class Message implements Serializable {
 		//   byte[]ID
 		//   int   value length
 		//   byte[]value
-		b.putInt(m.getInstance());
+		b.putLong(m.getInstance());
 		b.putInt(m.getSender());
 		b.putShort((short)m.getReceiver().getId());
 		b.putShort((short)m.getType().getId());
